revise PromoTypes
