+++ src/objects/Card/values/Finish.ts
@@ -1,0 +1,7 @@
+export enum ScryfallFinish {
+  Nonfoil = "nonfoil",
+  Foil = "foil",
+  Etched = "etched",
+}
+
+export type ScryfallFinishLike = ScryfallFinish | `${ScryfallFinish}`;
