+++ src/objects/Card/values/index.ts
@@ -1,6 +1,6 @@
 export * from "./BorderColor";
 export * from "./Color";
-export * from "./Finishes";
+export * from "./Finish";
 export * from "./Format";
 export * from "./Frame";
 export * from "./FrameEffect";
